refactor(read-projections): project street name read model in dedicated StreetNameReadProjection
Move the street name CRUD event handling (maintaining StreetNameReadItem) out of
RoadSegmentReadProjection into its own StreetNameReadProjection, mirroring OrganizationReadProjection.
RoadSegmentReadProjection keeps the street name event handlers but only to refresh the denormalized
labels on linked road segments via UpdateStreetNameLabels.
